pyproject.toml
[project]
name = "rapwinews-api"
version = "0.1.0"
description = "ФAPI for RAPSI mobile app"
authors = [
    {name = "Satkov",email = "g.satkov@mail.ru"}
]
license = {text = "Apache 2.0"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (==5.1)",
    "djangorestframework (>=3.15,<4.0)",
    "feedparser (>=6.0.11,<7.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
